requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
flake8==3.8.3
Flask==1.1.2
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.6.0
pyflakes==2.2.0
python-dotenv==0.14.0
requests==2.24.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.2.0
